refactor(lowy): extract placementPolicy module — separate volatility axes
Lowy's #844 fourth-pass review flagged \`resolvePlacementBucket\` +
\`getBucketFor\` as a future shared seam ("if a second tiling mode
arrives and its placeNew requires git-aware bucket resolution,
extract to canvas/placementPolicy.ts"). Action it now: the bucketing
rule is a different volatility axis from the packing algorithm
(\`repoIslands.ts\` — pure, deterministic, opaque to git) and from
canvas orchestration (\`useCanvasArrange.ts\` — write coordination,
recenter, mobile gate).

The new \`canvas/placementPolicy.ts\` is the third axis: "what counts
as the same group for placement", today \`terminalKey().group\` with a
sibling-cwd fallback. Future tiling modes that need git-aware
per-create placement reuse \`getBucketFor\` and
\`resolvePlacementBucket\` directly without dragging in the rest of
useCanvasArrange.

Both functions take \`store\` as their first argument so the module
itself stays free of hook-instance coupling — they're plain functions,
testable without a SolidJS owner.
